dev-requirements.txt
pytest
pytest-cov
pytest-sugar
pytest-django
